+++ include/dynamics/CR3BP.hpp
@@ -5,17 +5,17 @@ namespace Dynamics {
 
 class CR3BPModel : public ODE {
 public:
-    CR3BPModel(float mu_);
+    CR3BPModel(double mu_);
 
     void derivatives(
-        const std::vector<Eigen::Vector3f>& positions,
-        const std::vector<Eigen::Vector3f>& velocities,
-        std::vector<Eigen::Vector3f>& dpos_dt,
-        std::vector<Eigen::Vector3f>& dvel_dt
+        const std::vector<Eigen::Vector3d>& positions,
+        const std::vector<Eigen::Vector3d>& velocities,
+        std::vector<Eigen::Vector3d>& dpos_dt,
+        std::vector<Eigen::Vector3d>& dvel_dt
     ) override;
 
 private:
-    float mu;
+    double mu;
 };
 
 }
