+++ include/dynamics/Euler.hpp
@@ -5,7 +5,7 @@ namespace Dynamics {
 
 class Euler : public Integrator {
 public:
-    void step(std::vector<Body>& bodies, ODE& model, float dt) override;
+    void step(std::vector<Body>& bodies, ODE& model, double dt) override;
 };
 
 }
