+++ include/dynamics/RK4.hpp
@@ -5,7 +5,7 @@ namespace Dynamics {
 
 class RK4 : public Integrator {
 public:
-    void step(std::vector<Body>& bodies, ODE& model, float dt) override;
+    void step(std::vector<Body>& bodies, ODE& model, double dt) override;
 };
 
 }
