+++ src/dynamics/Euler.cpp
@@ -2,17 +2,17 @@
 
 namespace Dynamics {
 
-void Euler::step(std::vector<Body>& bodies, ODE& model, float dt) {
+void Euler::step(std::vector<Body>& bodies, ODE& model, double dt) {
 
     int n = bodies.size();
 
-    std::vector<Eigen::Vector3f> x(n), v(n);
+    std::vector<Eigen::Vector3d> x(n), v(n);
     for (int i = 0; i < n; i++) {
         x[i] = bodies[i].getPosition();
         v[i] = bodies[i].getVelocity();
     }
 
-    std::vector<Eigen::Vector3f> dx, dv;
+    std::vector<Eigen::Vector3d> dx, dv;
     model.derivatives(x, v, dx, dv);
 
     for (int i = 0; i < n; i++) {
